Moved NNParameterInversion to its own file.
Addresses #8
